restructure graph home
