+++ packages/model/addon/-private/references/has-many.ts
@@ -9,7 +9,7 @@ import { resolve } from 'rsvp';
 import { ManyArray } from 'ember-data/-private';
 
 import { DEPRECATE_PROMISE_PROXIES } from '@ember-data/private-build-infra/deprecations';
-import type { Graph } from '@ember-data/record-data/-private/graph';
+import type { Graph } from '@ember-data/record-data/-private/graph/graph';
 import type ManyRelationship from '@ember-data/record-data/-private/relationships/state/has-many';
 import type Store from '@ember-data/store';
 import { recordIdentifierFor } from '@ember-data/store';
