+++ packages/record-data/addon/-private/graph/operations/remove-from-related-records.ts
@@ -5,7 +5,7 @@ import type { StableRecordIdentifier } from '@ember-data/types/q/identifier';
 import type ManyRelationship from '../../relationships/state/has-many';
 import type { RemoveFromRelatedRecordsOperation } from '../-operations';
 import { isHasMany, notifyChange } from '../-utils';
-import type { Graph } from '../index';
+import type { Graph } from '../graph';
 import { removeFromInverse } from './replace-related-records';
 
 export default function removeFromRelatedRecords(
