+++ packages/record-data/addon/-private/graph/operations/replace-related-records.ts
@@ -6,7 +6,7 @@ import type { StableRecordIdentifier } from '@ember-data/types/q/identifier';
 import type ManyRelationship from '../../relationships/state/has-many';
 import type { ReplaceRelatedRecordsOperation } from '../-operations';
 import { isBelongsTo, isHasMany, isNew, notifyChange } from '../-utils';
-import type { Graph } from '../index';
+import type { Graph } from '../graph';
 
 /*
     case many:1
